Add extensions submodule, change imports and the like.
