+++ rdw_ingestion_tools/api/turn_bq/requests/cards.py
@@ -1,8 +1,9 @@
 from attrs import define
 from httpx import Client
-from pandas import DataFrame, concat, json_normalize
+from pandas import DataFrame
 
-from .. import get_paginated
+from ..extensions.dataframe import concatenate
+from ..extensions.httpx import get_paginated
 
 
 @define
@@ -18,12 +19,7 @@ class Cards:
 
         cards_generator = get_paginated(self.client, url)
 
-        try:
-            cards = concat(
-                [json_normalize(obj, sep="_") for obj in cards_generator]
-            )
-        except ValueError:
-            cards = DataFrame()
+        cards = concatenate(cards_generator)
 
         return cards
 
@@ -34,11 +30,6 @@ class Cards:
 
         cards_generator = get_paginated(self.client, url)
 
-        try:
-            cards = concat(
-                [json_normalize(obj, sep="_") for obj in cards_generator]
-            )
-        except ValueError:
-            cards = DataFrame()
+        cards = concatenate(cards_generator)
 
         return cards
