+++ rdw_ingestion_tools/api/turn_bq/requests/flow_results.py
@@ -1,8 +1,9 @@
 from attrs import define
 from httpx import Client
-from pandas import DataFrame, concat, json_normalize
+from pandas import DataFrame
 
-from .. import get_paginated
+from ..extensions.dataframe import concatenate
+from ..extensions.httpx import get_paginated
 
 
 @define
@@ -29,15 +30,7 @@ class FlowResults:
             url,
         )
 
-        try:
-            flow_results = concat(
-                [
-                    json_normalize(obj, sep="_")
-                    for obj in flow_results_generator
-                ]
-            )
-        except ValueError:
-            flow_results = DataFrame()
+        flow_results = concatenate(flow_results_generator)
 
         return flow_results
 
@@ -61,14 +54,6 @@ class FlowResults:
             self.client, url, page_size=100, **params
         )
 
-        try:
-            flow_results = concat(
-                [
-                    json_normalize(obj, sep="_")
-                    for obj in flow_results_generator
-                ]
-            )
-        except ValueError:
-            flow_results = DataFrame()
+        flow_results = concatenate(flow_results_generator)
 
         return flow_results
