+++ rdw_ingestion_tools/api/turn_bq/requests/flow_results_data_packages.py
@@ -1,8 +1,9 @@
 from attrs import define
 from httpx import Client
-from pandas import DataFrame, concat, json_normalize
+from pandas import DataFrame
 
-from .. import get_paginated
+from ..extensions.dataframe import concatenate
+from ..extensions.httpx import get_paginated
 
 
 @define
@@ -31,15 +32,9 @@ class FlowResultsDataPackages:
             url,
         )
 
-        try:
-            flow_results_data_packages = concat(
-                [
-                    json_normalize(obj, sep="_")
-                    for obj in flow_results_data_packages_generator
-                ]
-            )
-        except ValueError:
-            flow_results_data_packages = DataFrame()
+        flow_results_data_packages = concatenate(
+            flow_results_data_packages_generator
+        )
 
         return flow_results_data_packages
 
@@ -63,14 +58,8 @@ class FlowResultsDataPackages:
             self.client, url, page_size=100, **params
         )
 
-        try:
-            flow_results_data_packages = concat(
-                [
-                    json_normalize(obj, sep="_")
-                    for obj in flow_results_data_packages_generator
-                ]
-            )
-        except ValueError:
-            flow_results_data_packages = DataFrame()
+        flow_results_data_packages = concatenate(
+            flow_results_data_packages_generator
+        )
 
         return flow_results_data_packages
