+++ rdw_ingestion_tools/api/turn_bq/requests/messages.py
@@ -1,8 +1,9 @@
 from attrs import define
 from httpx import Client
-from pandas import DataFrame, concat, json_normalize
+from pandas import DataFrame
 
-from .. import get_paginated
+from ..extensions.dataframe import concatenate
+from ..extensions.httpx import get_paginated
 
 
 @define
@@ -26,11 +27,6 @@ class Messages:
             self.client, url, page_size=100, **params
         )
 
-        try:
-            messages = concat(
-                [json_normalize(obj, sep="_") for obj in messages_generator]
-            )
-        except ValueError:
-            messages = DataFrame()
+        messages = concatenate(messages_generator)
 
         return messages
