+++ rdw_ingestion_tools/api/turn_bq/requests/statuses.py
@@ -1,8 +1,9 @@
 from attrs import define
 from httpx import Client
-from pandas import DataFrame, concat, json_normalize
+from pandas import DataFrame
 
-from .. import get_paginated
+from ..extensions.dataframe import concatenate
+from ..extensions.httpx import get_paginated
 
 
 @define
@@ -26,11 +27,6 @@ class Statuses:
             self.client, url, page_size=100, **params
         )
 
-        try:
-            status = concat(
-                [json_normalize(obj, sep="_") for obj in status_generator]
-            )
-        except ValueError:
-            status = DataFrame()
+        status = concatenate(status_generator)
 
         return status
